# cal_track_ai/cal_track_ai/server/src/routes/meals.ts
import { Request, Response } from "express";
import { randomUUID } from "crypto";
import { prisma } from "../utils/prisma";
import {
  computeAccuracyMetrics,
  mealAnalyzeSchema,
  mealSaveSchema,
  mealUpdateSchema,
  validateMealEstimate,
} from "../utils/validation";
import { authMiddleware } from "../middleware/auth";
import { config, logger } from "../utils/config";
import { getAiAnalysisStatus } from "../utils/systemSettings";

export async function analyzeMeal(req: Request, res: Response) {
  const parseResult = mealAnalyzeSchema.safeParse(req.body);
  if (!parseResult.success) {
    return res.status(400).json({ detail: parseResult.error.errors });
  }

  // Human-oversight kill-switch (Art. 14(2)): lets an admin stop the AI
  // meal-analysis pipeline (e.g. mid-incident) without a deploy. Checked
  // before the model call so a "stopped" system makes zero AI requests.
  const oversightStatus = await getAiAnalysisStatus();
  if (!oversightStatus.enabled) {
    logger.warn(`meal_analysis_blocked_by_kill_switch user_id=${req.user?.id}`);
    return res.status(503).json({
      detail:
        "AI meal analysis has been paused by an administrator and is temporarily unavailable. You can still log meals manually.",
      ai_analysis_enabled: false,
    });
  }

  if (!config.geminiApiKey) {
    return res.status(500).json({ detail: "AI service not configured" });
  }

  const { image_base64, meal_type } = parseResult.data;

  try {
    const prompt = `You are a nutrition analysis AI. Given a food image, return ONLY a valid JSON
object (no markdown fences) with these fields:
{
  "is_food": boolean,
  "confidence": number,            // 0.0–1.0, your confidence in this estimate
  "food_name": "Name of the dish/food",
  "calories": number,
  "protein": number, "carbs": number, "fats": number, "fiber": number,
  "health_score": number,          // 1–10
  "health_score_rationale": "1–2 sentences explaining HOW the health score was derived",
  "quantity_grams": number,
  "ingredients": ["..."],
  "meal_description": "Brief description",
  "assumptions": "Any portion/ingredient assumptions made"
}
Base health_score on nutrient density, processing level, sugar/sodium, and portion.
Always explain the health_score in health_score_rationale.`;

// Be as accurate as possible with portion estimation. Always return valid JSON only.`;

    let mimeType = "image/jpeg";
    let imageData = image_base64.trim();
    if (imageData.startsWith("data:")) {
      const parts = imageData.split(",", 2);
      imageData = parts[1];
      if (parts[0].includes(";base64")) {
        mimeType = parts[0].slice(5).split(";")[0];
      }
    }

    const response = await fetch(
      `https://generativelanguage.googleapis.com/v1beta/models/${config.geminiModel}:generateContent?key=${config.geminiApiKey}`,
      {
        method: "POST",
        headers: { "Content-Type": "application/json" },
        body: JSON.stringify({
          contents: [
            {
              parts: [
                { text: prompt },
                {
                  inlineData: {
                    mimeType,
                    data: imageData,
                  },
                },
                { text: "Analyze this food image and provide the nutritional information as JSON." },
              ],
            },
          ],
        }),
      }
    );

    const result = await response.json();
    logger.info(`Gemini raw response: ${JSON.stringify(result, null, 2)}`);

    const responseText =
      (result as any).candidates?.[0]?.content?.parts?.[0]?.text || "";
    logger.info(`Gemini response: ${responseText}`);

    if (!responseText) {
      logger.error("Gemini response text is empty.");
      return res.status(500).json({ detail: "Analysis failed: Empty response from AI" });
    }

    const cleanText = responseText
      .replace(/^```json/, "")
      .replace(/^```/, "")
      .replace(/```$/, "")
      .trim();

    const AI_DISCLOSURE = {
  model_provider: "Google",
  model_name: config.geminiModel,
  analysis_type: "ai_estimate",
  data_provenance:
    "This system uses a third-party foundation model and does not train or curate " +
    "its own dataset. See data_governance_doc for the full provenance statement and " +
    "what we do/do not independently verify.",
  data_governance_doc: "https://github.com/kartikjain007/cal_ai/blob/main/docs/DATA_GOVERNANCE.md",
  limitations: [
    "Portion size and calorie/macro values are visual estimates and may differ meaningfully from a lab-measured value.",
    "Not validated for medical, clinical, or allergen-safety decisions — verify with a qualified professional before relying on this for a health condition.",
    "Accuracy may be lower for mixed dishes, unfamiliar cuisines, or partially obscured food.",
    "Not cross-checked against a certified nutrition database (e.g. USDA FoodData Central).",
  ],
  intended_use: "General wellness / self-tracking reference only — not a certified nutrition or medical measurement.",
};


    try {
      const nutritionData = JSON.parse(cleanText) as Record<string, unknown>;

      if (nutritionData.is_food === false) {
        logger.info(`Meal analysis refused: ${nutritionData.refusal_reason}`);
        return res.status(422).json({
          detail: nutritionData.refusal_reason || "This image doesn't appear to show food and was not analyzed.",
          is_food: false,
        });
      }

      const calories = Number(nutritionData.calories) || 0;
      const protein = Number(nutritionData.protein) || 0;
      const carbs = Number(nutritionData.carbs) || 0;
      const fats = Number(nutritionData.fats) || 0;
      const fiber = Number(nutritionData.fiber) || 0;
      const quantityGrams = Number(nutritionData.quantity_grams) || 100;
      const confidence = typeof nutritionData.confidence === "number" ? nutritionData.confidence : undefined;

      const quality = validateMealEstimate({
        calories,
        protein,
        carbs,
        fats,
        fiber,
        quantityGrams,
        confidence,
      });
      if (quality.needsReview) {
        logger.warn(`meal_estimate_flagged reasons=${quality.reasons.join(",")}`);
      }

      return res.json({
        food_name: nutritionData.food_name || "Unknown Food",
        calories,
        protein,
        carbs,
        fats,
        fiber,
        health_score: nutritionData.health_score || 5,
        quantity_grams: quantityGrams,
        ingredients: nutritionData.ingredients || [],
        meal_description: nutritionData.meal_description || "",
        meal_type,
        confidence,
        needs_review: quality.needsReview,
        flag_reasons: quality.reasons,
        ai_disclosure: AI_DISCLOSURE,
      });
    } catch (e) {
      logger.error(`Failed to parse Gemini response: ${e}`);
      return res.status(500).json({ detail: `Analysis failed: Invalid response from AI` });
    }
  } catch (error) {
    logger.error(`Meal analysis failed: ${error}`);
    if (error instanceof Error) {
      return res.status(500).json({ detail: `Analysis failed: ${error.message}` });
    }
    return res.status(500).json({ detail: `Analysis failed: An unknown error occurred` });
  }
}

export async function getTodaySummary(req: Request, res: Response) {
  const { date } = req.query;
  const userId = req.user!.id;

  let target: Date;
  let nextDay: Date;

  if (date) {
    try {
      target = new Date(date as string);
      target.setUTCHours(0, 0, 0, 0);
      nextDay = new Date(target);
      nextDay.setDate(nextDay.getDate() + 1);
    } catch {
      target = new Date();
      target.setUTCHours(0, 0, 0, 0);
      nextDay = new Date(target);
      nextDay.setDate(nextDay.getDate() + 1);
    }
  } else {
    target = new Date();
    target.setUTCHours(0, 0, 0, 0);
    nextDay = new Date(target);
    nextDay.setDate(nextDay.getDate() + 1);
  }

  const meals = await prisma.meal.findMany({
    where: {
      userId,
      loggedAt: {
        gte: target,
        lt: nextDay,
      },
    },
  });

  const totals = meals.reduce(
    (acc, m) => ({
      calories: (acc.calories || 0) + (m.calories || 0),
      protein: (acc.protein || 0) + (m.protein || 0),
      carbs: (acc.carbs || 0) + (m.carbs || 0),
      fats: (acc.fats || 0) + (m.fats || 0),
    }),
    {} as Record<string, number>
  );

  const flaggedMeals = meals.filter((m) => m.flaggedForReview);
  const requestId = req.requestId ?? randomUUID();
  const generatedAt = new Date().toISOString();

  if (flaggedMeals.length > 0) {
    logger.warn(
      `today_summary_includes_flagged_meals request_id=${requestId} user_id=${userId} flagged_count=${flaggedMeals.length} total_meals=${meals.length}`
    );
  }

  // Formal accuracy metrics (Art. 15.1) — computed from each contributing
  // meal's model-reported confidence, not a fixed label.
  const accuracyMetrics = computeAccuracyMetrics(meals);

  return res.json({
    total_calories: totals.calories || 0,
    total_protein: totals.protein || 0,
    total_carbs: totals.carbs || 0,
    total_fats: totals.fats || 0,
    meal_count: meals.length,
    goal_calories: req.user!.dailyCalories,
    goal_protein: req.user!.dailyProtein,
    goal_carbs: req.user!.dailyCarbs,
    goal_fats: req.user!.dailyFats,
    metadata: {
      request_id: requestId,
      generated_at: generatedAt,
    },
    data_quality: {
      source: "ai_estimated",
      method: "gemini_nutrition_analysis_per_meal",
      confidence_level: "approximate",
      flagged_meal_count: flaggedMeals.length,
      flagged_meal_ids: flaggedMeals.map((m) => String(m.id)),
      accuracy_metrics: accuracyMetrics,
      disclaimer:
        "Totals are aggregated from AI-estimated per-meal values, not lab-measured; individual meal estimates may deviate meaningfully from actual nutrient content. flagged_meal_count indicates how many contributing meals failed a plausibility check (see docs/DATA_GOVERNANCE.md) and should be reviewed before trusting this total.",
    },
    instructions_for_use: {
      intended_use:
        "Informational self-tracking reference comparing today's estimated intake to the user's personal goals. Not a medical, clinical, or regulatory nutrition measurement.",
      how_to_interpret: [
        "total_* fields are the sum of AI-estimated per-meal values for the queried day, not a lab-measured total.",
        "goal_* fields reflect the user's current profile settings and may be out of date if the user hasn't updated them recently.",
        "Treat total_* as directional guidance against goal_* (e.g. 'roughly on track'), not a precise deficit/surplus figure.",
        "If data_quality.flagged_meal_count > 0, this total includes at least one meal that failed a plausibility check — review data_quality.flagged_meal_ids via GET /api/meals/:mealId before relying on the total.",
        "data_quality.accuracy_metrics.avg_confidence is the mean of each contributing meal's AI self-reported confidence (0.0-1.0), not a measured error rate — see accuracy_metrics.basis for what it does and doesn't represent.",
        "meal_count reflects saved meals only; unsaved/discarded analyses are not included.",
      ],
      human_oversight:
        "Any contributing meal's estimated values can be corrected via PUT /api/meals/:mealId, which changes this total on the next request. There is no automated correction — a human must review and edit.",
      instructions_doc: "https://github.com/kartikjain007/cal_ai/blob/main/docs/DATA_GOVERNANCE.md",
    },
  });
}

export async function saveMeal(req: Request, res: Response) {
  const parseResult = mealSaveSchema.safeParse(req.body);
  if (!parseResult.success) {
    return res.status(400).json({ detail: parseResult.error.errors });
  }

  const userId = req.user!.id;
  const data = parseResult.data;

  const loggedAt = data.logged_at ? new Date(data.logged_at) : new Date();

  const quality = validateMealEstimate({
    calories: data.calories,
    protein: data.protein,
    carbs: data.carbs,
    fats: data.fats,
    fiber: data.fiber,
    quantityGrams: data.quantity_grams,
    confidence: data.confidence,
  });
  if (quality.needsReview) {
    logger.warn(`meal_save_flagged reasons=${quality.reasons.join(",")}`);
  }

  const meal = await prisma.meal.create({
    data: {
      userId,
      foodName: data.food_name,
      calories: data.calories,
      protein: data.protein,
      carbs: data.carbs,
      fats: data.fats,
      fiber: data.fiber,
      healthScore: data.health_score,
      quantityGrams: data.quantity_grams,
      ingredients: data.ingredients,
      mealDescription: data.meal_description,
      mealType: data.meal_type,
      imageBase64: data.image_base64,
      loggedAt,
      confidence: data.confidence,
      flaggedForReview: quality.needsReview,
    },
  });

  return res.json({
    id: String(meal.id),
    food_name: meal.foodName,
    calories: meal.calories,
    protein: meal.protein,
    carbs: meal.carbs,
    fats: meal.fats,
    fiber: meal.fiber,
    health_score: meal.healthScore,
    quantity_grams: meal.quantityGrams,
    ingredients: meal.ingredients,
    meal_description: meal.mealDescription,
    meal_type: meal.mealType,
    logged_at: meal.loggedAt?.toISOString(),
    confidence: meal.confidence,
    needs_review: meal.flaggedForReview,
    flag_reasons: quality.reasons,
  });
}

export async function getMeals(req: Request, res: Response) {
  const { date } = req.query;
  const userId = req.user!.id;

  let where: Record<string, unknown> = { userId };

  if (date) {
    try {
      const targetDate = new Date(date as string);
      targetDate.setUTCHours(0, 0, 0, 0);
      const nextDay = new Date(targetDate);
      nextDay.setDate(nextDay.getDate() + 1);

      where = {
        ...where,
        loggedAt: {
          gte: targetDate,
          lt: nextDay,
        },
      };
    } catch {
      // ignore invalid date
    }
  }

  // image_base64 is excluded here (list view) and only returned by
  // GET /api/meals/:mealId (detail view) — a list of up to 100 meals
  // shouldn't pull every full-resolution photo out of the DB and over the
  // wire just to render a card that doesn't show it at full size anyway.
  const meals = await prisma.meal.findMany({
    where,
    orderBy: { loggedAt: "desc" },
    take: 100,
    select: {
      id: true,
      foodName: true,
      calories: true,
      protein: true,
      carbs: true,
      fats: true,
      fiber: true,
      healthScore: true,
      quantityGrams: true,
      mealType: true,
      mealDescription: true,
      ingredients: true,
      loggedAt: true,
      confidence: true,
      flaggedForReview: true,
    },
  });

  // Formal data quality metrics (Art. 10.1) — each meal already carries its
  // own confidence/needs_review, but a list of individually-flagged records
  // doesn't by itself surface the quality of the *page as a whole*. Reuse
  // the same computeAccuracyMetrics() helper as today-summary/weekly so a
  // caller can see at a glance how reliable this batch of estimates is.
  const requestId = req.requestId ?? randomUUID();
  const flaggedCount = meals.filter((m) => m.flaggedForReview).length;
  logger.info(
    `meals_list_view request_id=${requestId} user_id=${userId} count=${meals.length} flagged_count=${flaggedCount}`
  );

  return res.json({
    meals: meals.map((m) => ({
      id: String(m.id),
      food_name: m.foodName,
      calories: m.calories,
      protein: m.protein,
      carbs: m.carbs,
      fats: m.fats,
      fiber: m.fiber,
      health_score: m.healthScore,
      quantity_grams: m.quantityGrams,
      meal_type: m.mealType,
      meal_description: m.mealDescription,
      ingredients: m.ingredients,
      logged_at: m.loggedAt?.toISOString(),
      confidence: m.confidence,
      needs_review: m.flaggedForReview,
    })),
    data_quality: {
      source: "ai_estimated",
      method: "gemini_nutrition_analysis_per_meal",
      accuracy_metrics: computeAccuracyMetrics(meals),
      flagged_meal_count: flaggedCount,
      // Art. 10.1 — training/validation/testing data quality. There is no
      // in-house dataset behind these estimates to report quality,
      // representativeness, or validation metrics on; this states that
      // scope explicitly rather than the gap being silent about it.
      training_data_provenance: {
        own_training_data: false,
        model_provider: "Google",
        model_name: config.geminiModel,
        statement:
          "Nutrition values are produced by a third-party foundation model (Google Gemini). This system does not train, fine-tune, or curate a training/validation/testing dataset of its own, so there is no in-house dataset whose quality, representativeness, or validation split this system can report metrics on — that obligation sits with the model provider, not this application. The data quality controls this system does apply and does control are: input validation before an image is sent to the model, and output plausibility checks after (see accuracy_metrics above, plus confidence thresholding, calorie-density, and macro-mismatch checks in docs/DATA_GOVERNANCE.md).",
        data_governance_doc: "https://github.com/kartikjain007/cal_ai/blob/main/docs/DATA_GOVERNANCE.md",
      },
      disclaimer:
        "Each meal is an AI-estimated value, not lab-measured; see accuracy_metrics.basis and docs/DATA_GOVERNANCE.md for methodology.",
    },
    metadata: {
      request_id: requestId,
      generated_at: new Date().toISOString(),
    },
  });
}

export async function getMealDetail(req: Request, res: Response) {
  const { mealId } = req.params;
  const userId = req.user!.id;

  const mId = parseInt(mealId, 10);
  if (isNaN(mId)) {
    return res.status(400).json({ detail: "Invalid meal ID format" });
  }

  const meal = await prisma.meal.findFirst({
    where: { id: mId, userId },
  });

  if (!meal) {
    return res.status(404).json({ detail: "Meal not found" });
  }

  return res.json({
    id: String(meal.id),
    food_name: meal.foodName,
    calories: meal.calories,
    protein: meal.protein,
    carbs: meal.carbs,
    fats: meal.fats,
    fiber: meal.fiber,
    health_score: meal.healthScore,
    quantity_grams: meal.quantityGrams,
    meal_type: meal.mealType,
    meal_description: meal.mealDescription,
    ingredients: meal.ingredients,
    image_base64: meal.imageBase64,
    logged_at: meal.loggedAt?.toISOString(),
    created_at: meal.createdAt.toISOString(),
    confidence: meal.confidence,
    needs_review: meal.flaggedForReview,
  });
}

export async function updateMeal(req: Request, res: Response) {
  const { mealId } = req.params;
  const userId = req.user!.id;

  const mId = parseInt(mealId, 10);
  if (isNaN(mId)) {
    return res.status(400).json({ detail: "Invalid meal ID format" });
  }

  const parseResult = mealUpdateSchema.safeParse(req.body);
  if (!parseResult.success) {
    return res.status(400).json({ detail: parseResult.error.errors });
  }

  const data = parseResult.data;

  const existing = await prisma.meal.findFirst({
    where: { id: mId, userId },
  });

  if (!existing) {
    return res.status(404).json({ detail: "Meal not found" });
  }

  const updateData: Record<string, unknown> = {};
  if (data.food_name !== undefined) updateData.foodName = data.food_name;
  if (data.quantity_grams !== undefined) updateData.quantityGrams = data.quantity_grams;
  if (data.calories !== undefined) updateData.calories = data.calories;
  if (data.protein !== undefined) updateData.protein = data.protein;
  if (data.carbs !== undefined) updateData.carbs = data.carbs;
  if (data.fats !== undefined) updateData.fats = data.fats;
  if (data.fiber !== undefined) updateData.fiber = data.fiber;

  await prisma.meal.update({
    where: { id: mId },
    data: updateData,
  });

  return res.json({ message: "Meal updated", ...updateData });
}

export async function deleteMeal(req: Request, res: Response) {
  const { mealId } = req.params;
  const userId = req.user!.id;

  const mId = parseInt(mealId, 10);
  if (isNaN(mId)) {
    return res.status(400).json({ detail: "Invalid meal ID format" });
  }

  const existing = await prisma.meal.findFirst({
    where: { id: mId, userId },
  });

  if (!existing) {
    return res.status(404).json({ detail: "Meal not found" });
  }

  await prisma.meal.delete({
    where: { id: mId },
  });

  return res.json({ message: "Meal deleted" });
}

export function registerMealRoutes(app: import("express").Express) {
  app.post("/api/meals/analyze", authMiddleware, analyzeMeal);
  app.get("/api/meals/today-summary", authMiddleware, getTodaySummary);
  app.post("/api/meals", authMiddleware, saveMeal);
  app.get("/api/meals", authMiddleware, getMeals);
  app.get("/api/meals/:mealId", authMiddleware, getMealDetail);
  app.put("/api/meals/:mealId", authMiddleware, updateMeal);
  app.delete("/api/meals/:mealId", authMiddleware, deleteMeal);
}